Update the Connection trait interface
Update the Connection trait interface to support a way for Connection trait
implementors to provide context-specific information needed to establish a
connection. An example of this is a database connection where the
context-specific information needed might include a username, a password, or a
database name. This removes the new function from the Connection trait and
instead requires a function used to initialize a connection be provided to the
ConnectionPool when it is created.