Cueball should terminate decoherence check on pool stop
Switch from using the periodic crate to the timer crate to manage the
decoherence work. This change makes it so that when the connection pool is
explicitly stopped or the connection pool goes out of scope then the threads
managing the periodic decoherence work also are terminated.

This is a backport to the 0.1 branch of the same change applied to the master
branch.