Introduce health checks on connections to rust-cueball. (#17)
Add support for connection health checks to cueball.

Update the Connection trait with methods that allow  trait implementors to provide connection state information to cueball. Implement re-tries and a connection-level checks using these methods to maintain a healthy connection pool. Provide default implementation so as not to break existing implementations that have not yet provided health methods. Bump the minor version of cueball to reflect the trait change.

Co-authored-by: Kelly McLaughlin <kelly.mclaughlin@joyent.com>